Merge in KI_ATLAS/atlas-versioning from feature/ATLAS-scenario-merge-in-the-middle to master

* commit 'bef754c8626ff6239782c2b27abf1dd5d9124f76':
  Add mergee scenario tests
  ATLAS-versioning: add new merge scenario

+++ src/main/java/ch/sbb/atlas/versioning/service/VersionableServiceImpl.java
@@ -81,7 +81,9 @@ public class VersionableServiceImpl implements VersionableService {
       }
       if (DELETE.equals(versionedObject.getAction())) {
         log(versionedObject);
-        deleteById.accept(versionedObject.getEntity().getId());
+        if(versionedObject.getEntity().getId() != null){
+          deleteById.accept(versionedObject.getEntity().getId());
+        }
       }
     }
   }
